requirements.txt
python-telegram-bot==20.3
aiohttp==3.8.5
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3
configparser==5.3.0
python-telegram-bot[job-queue]